V17-8
28/03/2020 V17-8 :
fix : GFX packs not found msg after an update
fix : third party software were not launched on the very first run of a game
fix : empty transferable cache detection failure on the very first run of a game
fix : lower size threshold of the new feature
fix : set some choices timeout in setup.bat 
fix : when creating new Graphic packs, setting the complete titleId list (regions) in rules.txt was broken
add : more scale directives in GFX packs V3 created

+++ tools/updateBatchFw.bat
@@ -233,7 +233,7 @@ REM : functions
 
         REM : delete old GFX packs archive
         set "gfxRar="!BFW_RESOURCES_PATH:"=!\V3_GFX_Packs.rar""
-        del /F /S !gfxRar! > NUL 2>&1
+        if exist !gfxRar! del /F /S !gfxRar! > NUL 2>&1
 
         REM : update fixBrokenSHortcuts in ALL shortcuts folder
         for /F "tokens=2 delims=~=" %%i in ('type !logFile! ^| find "Create " 2^>NUL') do (
@@ -250,40 +250,9 @@ REM : functions
             )
         )
 
-        REM : V15-8 clean DESIRED_ASPECT_RATIO in all host log
-        REM : search in all Host_*.log
-        set "pat="!BFW_PATH:"=!\logs\Host_*.log""
-
-        for /F "delims=~" %%i in ('dir /S /B !pat! 2^>NUL') do (
-            set "currentLogFile="%%i""
-
-            call:cleanArFromHostLogFile !currentLogFile!
-        )
-
     goto:eof
     REM : ------------------------------------------------------------------
 
-    :cleanArFromHostLogFile
-        REM : pattern to ignore in log file
-        set "log="%~1""
-
-        set "logFileTmp="!log:"=!.bfw_tmp""
-        if exist !logFileTmp! (
-            del /F !log! > NUL 2>&1
-            move /Y !logFileTmp! !log! > NUL 2>&1
-        )
-        if exist !logFileTmp! (
-            del /F !log! > NUL 2>&1
-            move /Y !logFileTmp! !log! > NUL 2>&1
-        )
-
-        type !log! | find /I /V "DESIRED_ASPECT_RATIO" > !logFileTmp!
-
-        del /F /S !log! > NUL 2>&1
-        move /Y !logFileTmp! !log! > NUL 2>&1
-
-    goto:eof
-    REM : ------------------------------------------------------------------
 
     REM : COMPARE VERSIONS : function to count occurences of a separator
     :countSeparators
